+++ src/hooks/queryHooks.ts
@@ -1,9 +1,6 @@
-import { IpostCommentParams, IputPostBody } from "@/_redux/slices/postSlices/getPostSlice";
-import { IComment, IPost, IUser } from "@/api/_types/apiModels";
-import { deleteApiJWT, getApi, getApiJWT, postApiJWT, putApiJWT } from "@/api/apis";
-import { createNotification } from "@/api/createNotification";
-import { createFormData } from "@/utils/createFormData";
-import { useMutation, useQuery, useQueryClient, useSuspenseQuery } from "@tanstack/react-query";
+import {  IUser } from "@/api/_types/apiModels";
+import {  getApi, getApiJWT } from "@/api/apis";
+import {  useQuery, useSuspenseQuery } from "@tanstack/react-query";
 // import { AxiosResponse } from "axios";
 
 // uesUserInfo에서 이름변경
@@ -19,24 +16,24 @@ export const useAuthUser = () => {
   });
 };
 
-export const useUnscheduledCards = (unscheduledChannelId: string) => {
-  return useSuspenseQuery({
-    queryKey: ['unscheduledCards'],
-    queryFn: async ()=>{
-      return await getApi<IPost[]>(`/posts/channel/${unscheduledChannelId}`)
-    },
-  })
-}
+// export const useUnscheduledCards = (unscheduledChannelId: string) => {
+//   return useSuspenseQuery({
+//     queryKey: ['unscheduledCards'],
+//     queryFn: async ()=>{
+//       return await getApi<IPost[]>(`/posts/channel/${unscheduledChannelId}`)
+//     },
+//   })
+// }
 
-export const useMyJoinCards = (id: string | undefined) => {
-  if(!id) console.error('user 정보가 없습니다!');
-  return useQuery({
-    queryKey: ['myJoinCards'],
-    queryFn: async ()=> {
-      return await getApi(`/users/${id}`)
-    },
-  })
-}
+// export const useMyJoinCards = (id: string | undefined) => {
+//   if(!id) console.error('user 정보가 없습니다!');
+//   return useQuery({
+//     queryKey: ['myJoinCards'],
+//     queryFn: async ()=> {
+//       return await getApi(`/users/${id}`)
+//     },
+//   })
+// }
 
 export const usePostsAuthor = <T>(userId: string | undefined) => {
   if(!userId) console.error('user 정보가 없습니다!');
@@ -61,25 +58,25 @@ export const useUsersInfo = <T>(userId: string | undefined) => {
   })
 }
 
-export const usePostDetail= <T>(postId: string) => {
-  if(!postId) console.error('postId 정보가 없습니다!');
-  return useSuspenseQuery({
-    queryKey: [`posts/${postId}`, postId],
-    queryFn: async ()=> {
-      console.log("usePostDetail 쿼리캐싱")
+// export const useGetPostDetail= <T>(postId: string) => {
+//   if(!postId) console.error('postId 정보가 없습니다!');
+//   return useSuspenseQuery({
+//     queryKey: [`posts/${postId}`, postId],
+//     queryFn: async ()=> {
+//       console.log("useGetPostDetail 쿼리캐싱")
 
-      return await getApi<T>(`/posts/${postId}`)
-    },
-    staleTime: Infinity,
-  })
-}
+//       return await getApi<T>(`/posts/${postId}`)
+//     },
+//     staleTime: Infinity,
+//   })
+// }
 
-// export const usePostDetail2= <IPost>(postId: string) => {
+// export const useGetPostDetail2= <IPost>(postId: string) => {
 //   if(!postId) console.error('postId 정보가 없습니다!');
 //   return useSuspenseQuery<AxiosResponse<IPost>, Error,string>({
 //     queryKey: [`posts/${postId}`, postId],
 //     queryFn: async ()=> {
-//       console.log("usePostDetail 쿼리캐싱")
+//       console.log("useGetPostDetail 쿼리캐싱")
 
 //       return await getApi<IPost>(`/posts/${postId}`)
 //     },
@@ -91,50 +88,50 @@ export const usePostDetail= <T>(postId: string) => {
 //   })
 // }
 
-export const usePutPostDetail = (postId?: string) =>{
-  const queryClient = useQueryClient();
-  const { mutate } = useMutation({
-    mutationFn: async (body: IputPostBody) => await  putApiJWT<IPost, FormData>(
-      `/posts/update`,
-      createFormData(body),
-    ),
-    onSuccess: async () =>{
-      await queryClient.invalidateQueries({
-        queryKey: [`posts/${postId}`, postId],
-      });
-    }
-  })
-  return {mutate}
-}
+// export const usePutPostDetail = (postId?: string) =>{
+//   const queryClient = useQueryClient();
+//   const { mutate } = useMutation({
+//     mutationFn: async (body: IputPostBody) => await putApiJWT<IPost, FormData>(
+//       `/posts/update`,
+//       createFormData(body),
+//     ),
+//     onSuccess: async () =>{
+//       await queryClient.invalidateQueries({
+//         queryKey: [`posts/${postId}`, postId],
+//       });
+//     }
+//   })
+//   return {mutate}
+// }
 
-export const usePostComment = ({ postId, postAuthorId } : IpostCommentParams) => {
-  const queryClient = useQueryClient();
-  const {mutate} = useMutation({
-    mutationFn : async (comment:string) => await postApiJWT<IComment>('/comments/create', {comment, postId}),
-    onSuccess: async (data) => {
-      createNotification({
-        notificationType: 'COMMENT',
-        notificationTypeId: data.data._id,
-        userId: postAuthorId,
-        postId,
-      });
-      await queryClient.invalidateQueries({
-        queryKey: [`posts/${postId}`, postId],
-      });
-    }
-  });
-  return { mutate }
-}
+// export const usePostComment = ({ postId, postAuthorId } : IpostCommentParams) => {
+//   const queryClient = useQueryClient();
+//   const {mutate} = useMutation({
+//     mutationFn : async (comment:string) => await postApiJWT<IComment>('/comments/create', {comment, postId}),
+//     onSuccess: async (data) => {
+//       createNotification({
+//         notificationType: 'COMMENT',
+//         notificationTypeId: data.data._id,
+//         userId: postAuthorId,
+//         postId,
+//       });
+//       await queryClient.invalidateQueries({
+//         queryKey: [`posts/${postId}`, postId],
+//       });
+//     }
+//   });
+//   return { mutate }
+// }
 
-export const useDeleteComment = (postId: string) =>{
-  const queryClient = useQueryClient();
-  const { mutate } = useMutation({
-    mutationFn: async (commentId:string) => await deleteApiJWT<IComment>('/comments/delete', {id: commentId}),
-    onSuccess: async () =>{
-      await queryClient.invalidateQueries({
-        queryKey: [`posts/${postId}`, postId],
-      });
-    }
-  })
-  return {mutate}
-}
+// export const useDeleteComment = (postId: string) =>{
+//   const queryClient = useQueryClient();
+//   const { mutate } = useMutation({
+//     mutationFn: async (commentId:string) => await deleteApiJWT<IComment>('/comments/delete', {id: commentId}),
+//     onSuccess: async () =>{
+//       await queryClient.invalidateQueries({
+//         queryKey: [`posts/${postId}`, postId],
+//       });
+//     }
+//   })
+//   return {mutate}
+// }
