+++ src/pages/DetailPage/DetailPost/PostContents.tsx
@@ -1,13 +1,13 @@
 import styled from '@emotion/styled';
 import { IPost, IPostTitleCustom } from '@/api/_types/apiModels';
-import { usePostDetail } from '@/hooks/queryHooks';
+import { useGetPostDetail } from '@/hooks/query/usePost';
 
 type PostContentsType = {
   postId: string;
 };
 
 export const PostContents = ({ postId }: PostContentsType) => {
-  const {data} = usePostDetail<IPost>(postId);
+  const {data} = useGetPostDetail<IPost>(postId);
   const {image, title}=data.data
   const responseTitle = JSON.parse(title) as IPostTitleCustom;
